Добавить системные сервисы ProxySQL и ProxyWeb с преднастройкой и генерацией учётных данных

+++ src/Command/ProjectDownCommand.php
@@ -14,6 +14,7 @@ use DockerCli\Project\DedicatedDatabaseComposeRenderer;
 use DockerCli\Project\DedicatedDatabaseDirectoryRemover;
 use DockerCli\Project\OpenRestyHostRenderer;
 use DockerCli\Project\ProjectRegistry;
+use DockerCli\Project\ProxySqlConfigurationSynchronizer;
 
 use function DockerCli\Util\join_path;
 
@@ -194,6 +195,10 @@ final class ProjectDownCommand extends AbstractCommand
             $this->removeDirectory($projectDirectory);
         }
         (new DedicatedDatabaseComposeRenderer())->render();
+        $proxySqlCode = (new ProxySqlConfigurationSynchronizer())->synchronize($output);
+        if ($proxySqlCode !== Command::SUCCESS) {
+            return $proxySqlCode;
+        }
 
         if ($input->getOption("erase")) {
             $this->removeDirectory($metadataDirectory);
